refactor(authentication): [IOPID-2907] Move remaining files related to authentication feature (#6881)
## Short description
This PR moves the remaining files related to the `authentication`
feature under the `ts/features/authentication` folder

## List of changes proposed in this pull request
- Moved all the remaining files that belong to the authentication area
under `ts/features/authentication` folder
- Updated all related imports

## How to test
No logic has been changed, but make sure all unit tests pass and that
the changes to imports haven’t introduced any regressions (simply by
running the command `yarn tsc:noemit`)
