+++ ts/features/authentication/common/saga/watchSessionExpiredSaga.ts
@@ -1,8 +1,8 @@
 import { takeLatest } from "typed-redux-saga/macro";
 import { getType } from "typesafe-actions";
-import { sessionExpired } from "../../features/authentication/common/store/actions";
-import { ReduxSagaEffect } from "../../types/utils";
-import { restartCleanApplication } from "../commons";
+import { sessionExpired } from "../store/actions";
+import { ReduxSagaEffect } from "../../../../types/utils";
+import { restartCleanApplication } from "../../../../sagas/commons";
 
 /**
  * Handles the expiration of the session while the user is using the app.
