+++ ts/features/authentication/login/cie/__tests__/ActivateNfcScreen.test.tsx
@@ -3,7 +3,7 @@ import { createStore } from "redux";
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import { Alert } from "react-native";
 import I18n from "../../../../../i18n";
-import { openNFCSettings } from "../../../../../utils/cie";
+import { openNFCSettings } from "../utils/cie";
 import { appReducer } from "../../../../../store/reducers";
 import { applicationChangeState } from "../../../../../store/actions/application";
 import { renderScreenWithNavigationStoreContext } from "../../../../../utils/testWrapper";
@@ -12,7 +12,7 @@ import ActivateNfcScreen from "../screens/ActivateNfcScreen";
 import { AUTHENTICATION_ROUTES } from "../../../common/navigation/routes";
 
 // Mock the openNFCSettings function
-jest.mock("../../../../../utils/cie", () => ({
+jest.mock("../utils/cie", () => ({
   openNFCSettings: jest.fn()
 }));
