+++ ts/features/authentication/login/cie/components/CieRequestAuthenticationOverlay.tsx
@@ -22,23 +22,22 @@ import * as T from "fp-ts/lib/Task";
 import * as TE from "fp-ts/lib/TaskEither";
 import { LoginUtilsError } from "@pagopa/io-react-native-login-utils";
 import CookieManager from "@react-native-cookies/cookies";
-import { IOColors } from "@pagopa/io-app-design-system";
-import { useHardwareBackButton } from "../../hooks/useHardwareBackButton";
-import I18n from "../../i18n";
-import { getIdpLoginUri } from "../../utils/login";
-import { closeInjectedScript } from "../../utils/webview";
-import { IOStyles } from "../core/variables/IOStyles";
-import { withLoadingSpinner } from "../helpers/withLoadingSpinner";
-import { lollipopKeyTagSelector } from "../../features/lollipop/store/reducers/lollipop";
-import { useIODispatch, useIOSelector } from "../../store/hooks";
-import { isMixpanelEnabled } from "../../store/reducers/persistedPreferences";
-import { regenerateKeyGetRedirectsAndVerifySaml } from "../../features/lollipop/utils/login";
-import { trackSpidLoginError } from "../../utils/analytics";
-import { isFastLoginEnabledSelector } from "../../features/authentication/fastLogin/store/selectors";
-import { isCieLoginUatEnabledSelector } from "../../features/authentication/login/cie/store/selectors";
-import { cieFlowForDevServerEnabled } from "../../features/authentication/login/cie/utils";
-import { selectedIdentityProviderSelector } from "../../features/authentication/common/store/selectors";
-import { OperationResultScreenContent } from "../screens/OperationResultScreenContent";
+import { IOColors, IOStyles } from "@pagopa/io-app-design-system";
+import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
+import I18n from "../../../../../i18n";
+import { getIdpLoginUri } from "../../../common/utils/login";
+import { closeInjectedScript } from "../../../../../utils/webview";
+import { withLoadingSpinner } from "../../../../../components/helpers/withLoadingSpinner";
+import { lollipopKeyTagSelector } from "../../../../../features/lollipop/store/reducers/lollipop";
+import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
+import { isMixpanelEnabled } from "../../../../../store/reducers/persistedPreferences";
+import { regenerateKeyGetRedirectsAndVerifySaml } from "../../../../../features/lollipop/utils/login";
+import { trackSpidLoginError } from "../../../../../utils/analytics";
+import { isFastLoginEnabledSelector } from "../../../fastLogin/store/selectors";
+import { isCieLoginUatEnabledSelector } from "../store/selectors";
+import { cieFlowForDevServerEnabled } from "../utils";
+import { selectedIdentityProviderSelector } from "../../../../../features/authentication/common/store/selectors";
+import { OperationResultScreenContent } from "../../../../../components/screens/OperationResultScreenContent";
 
 const styles = StyleSheet.create({
   errorContainer: {
