+++ ts/features/authentication/login/idp/screens/IdpLoginScreen.tsx
@@ -43,8 +43,8 @@ import {
   getIdpLoginUri,
   getIntentFallbackUrl,
   onLoginUriChanged
-} from "../../../../../utils/login";
-import { getSpidErrorCodeDescription } from "../../../../../utils/spidErrorCode";
+} from "../../../common/utils/login";
+import { getSpidErrorCodeDescription } from "../utils/spidErrorCode";
 import {
   assistanceToolRemoteConfig,
   handleSendAssistanceLog
