+++ ts/features/itwallet/identification/screens/ItwIdentificationIdpSelectionScreen.tsx
@@ -3,7 +3,7 @@ import { useFocusEffect } from "@react-navigation/native";
 import { useRef, useCallback } from "react";
 import { SpidIdp } from "../../../../../definitions/content/SpidIdp";
 import { isReady } from "../../../../common/model/RemoteValue";
-import IdpsGrid from "../../../../components/IdpsGrid";
+import IdpsGrid from "../../../authentication/login/idp/components/IdpsGrid";
 import { IOScrollViewWithLargeHeader } from "../../../../components/ui/IOScrollViewWithLargeHeader";
 import { randomOrderIdps } from "../../../authentication/login/idp/screens/IdpSelectionScreen";
 import { loadIdps } from "../../../../store/actions/content";
