+++ ts/features/itwallet/identification/screens/cie/ItwActivateNfcScreen.tsx
@@ -9,7 +9,7 @@ import { Alert } from "react-native";
 import { useFocusEffect } from "@react-navigation/native";
 import { IOScrollViewWithLargeHeader } from "../../../../../components/ui/IOScrollViewWithLargeHeader";
 import I18n from "../../../../../i18n";
-import * as cieUtils from "../../../../../utils/cie";
+import * as cieUtils from "../../../../authentication/login/cie/utils/cie";
 import { ItwEidIssuanceMachineContext } from "../../../machine/provider";
 import {
   trackItWalletCieNfcActivation,
