+++ ts/features/itwallet/identification/screens/spid/ItwSpidIdpLoginScreen.tsx
@@ -11,7 +11,7 @@ import {
 } from "../../../../../hooks/useHeaderSecondLevel";
 import I18n from "../../../../../i18n";
 import { originSchemasWhiteList } from "../../../../authentication/common/utils/originSchemasWhiteList";
-import { getIntentFallbackUrl } from "../../../../../utils/login";
+import { getIntentFallbackUrl } from "../../../../authentication/common/utils/login";
 import { useItwDismissalDialog } from "../../../common/hooks/useItwDismissalDialog";
 import {
   selectAuthUrlOption,
